backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "chromadb>=1.5.8",
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "joblib>=1.5.3",
    "openai>=2.32.0",
    "pandas>=3.0.2",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.4.1",
    "uvicorn[standard]>=0.45.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
